requirements.txt
Django=3.1
pytube3=9.6.4